livekit-agent/requirements copy.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==3.7.1
av==14.4.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
ctranslate2==4.6.0
diskcache==5.6.3
fastapi==0.104.1
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
gpt4all==2.0.2
h11==0.16.0
hf-xet==1.1.4
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
livekit==1.0.2
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx<3.0.0
numba==0.61.2
numpy<=2.2.0
onnxruntime==1.22.0
openai-whisper==20240930
packaging==25.0
protobuf==6.31.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pyttsx3==2.98
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
scipy==1.15.3
setuptools==80.9.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
starlette==0.27.0
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
tqdm==4.67.1
transformers==4.41.2
triton==2.2.0
types-protobuf==6.30.2.20250516
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.23.2

